ref: split the step definitions
