+++ test/support/world.ts
@@ -1,0 +1,5 @@
+export default class Context {
+  public app;
+  public response;
+  public preRequest;
+}
